Casted some key values to int to print out the value in the Hashmap

+++ src/College/Departement.java
@@ -28,8 +28,8 @@ public class Departement {
         return counter ;
     }
 
-    public String getDepartementNameById(int key){
-        return DepartementArray.get(key);
+    public String getDepartementNameById(byte key){
+        return DepartementArray.get((int)key);
     }
     public void addDepartement(){}
